resources and ultra basic AI

+++ Assets/Scripts/PlanetData.cs
@@ -29,7 +29,7 @@ public class PlanetData : MonoBehaviour
 		if (extractors.Count == 0)	return;
 
 		//give resources to all in the list, limited by how many resources there are
-		int rate = (int)Mathf.Min(extractionRate * extractors.Count, resources) / extractors.Count;
+		float rate = Mathf.Min(extractionRate * extractors.Count, resources) / extractors.Count;
 		foreach (ExtractorUnit unit in extractors) {
 			unit.GiveResources(rate);
 			resources -= rate;
